Bump Relay Versions
This marks an interesting point in development where the project can
be used to generate a blog.

Bumping the Relay versions solved an issue where Data Masking was
being applied incorrectly. Currently, using leo-plugin-blogpost with
core, one can display both individual post pages and "aggregate" pages
due to a `BlogPostConnection` on the root.