Enable Plugin Configuration
Due to switching to
[cosmiconfig](https://github.com/davidtheclark/cosmiconfig) `.leorc`
can now be a variety of things:

* a key in package.json, `{ leo: {} }`
* a JSON or YAML file at `.leorc`
* a `leo.config.js` file, which is a valid JS module (this allows for
  arbitrary JS code)
* In the future, a `--config` CLI arg can point to the config file

`rc` will be removed in a future commit

----

Let's modify the markdown plugin so that it uses our special
markdown-it instance. We want a raw version that doesn't have any
pesky plugins added, so it'll look very vanilla. Note that the config
key is the *same* as the name of the plugin.

```javascript
module.exports = {
  "plugins": [
    "@sa-labs/leo-plugin-blogpost",
    "@sa-labs/leo-plugin-markdown",
    "@sa-labs/leo-plugin-images"
  ],
  "@sa-labs/leo-plugin-markdown": {
      instance: require('markdown-it')({
          html: true,
          linkify: true,
          typographer: true
      })
  }
}
```

Plugin authors can use these options in their index.js file. For
plugin authors, this means they can offer customization options that
directly affect the webpack config. For example, in this case we use
the new markdown-it instance to modify the webpack config key
`@saLabs/leoPluginMarkdown/loader`. This key is used to pass options
to the markdown-loader. Effectively, this means plugin authors can
offer users configuration in `leo.config.js` that can be used to
modify webpack configuration.